+++ sunpy/io/setup_package.py
@@ -12,7 +12,7 @@ from astropy_helpers import setup_helpers
 
 def get_extensions():
 
-    if platform.system() == 'Windows' or sys.version_info.major == 3:
+    if platform.system() == 'Windows':
         return list()
     else:
         # 'numpy' will be replaced with the proper path to the numpy includes